+++ components/Chip/setup.ts
@@ -2,7 +2,9 @@ import { applyDefaults, cn, opt, partializeClasses } from 'utils';
 import { iComponent, iFwd } from 'types';
 
 export type iChip = iComponent<HTMLLIElement> & {
+  onRemove?: (ev: Event) => void;
   fwd: Partial<{
+    content: iFwd<HTMLSpanElement>;
     remove_button: iFwd<HTMLButtonElement>;
   }>;
 };
@@ -14,7 +16,7 @@ const defaults: iChip = {
 export default (props: Partial<iChip>) => {
   const p = applyDefaults<iChip>(defaults, props);
 
-  const { remove_button } = p.fwd;
+  const { remove_button, content } = p.fwd;
 
   const classes = partializeClasses({
     chip: opt(
@@ -26,6 +28,11 @@ export default (props: Partial<iChip>) => {
       cn('comp-chip_remove'),
       remove_button?.class,
       remove_button?.nostyle || p.nostyleAll,
+    ),
+    content: opt(
+      cn('comp-chip_content'),
+      content?.class,
+      content?.nostyle || p.nostyleAll,
     ),
   });