+++ routes/index.tsx
@@ -1,8 +1,13 @@
-// import { } from '../mod.ts';
+import Island from '../islands/Island.tsx';
+import { Header, Main } from '../mod.ts';
 
 export default function Home() {
   return (
     <div>
+      <Header></Header>
+      <Main layout_type='center'>
+        <Island />
+      </Main>
     </div>
   );
 }